+++ examples/ktor-server/src/main/kotlin/dev/nesk/akkurate/examples/ktor/server/plugins/Routing.kt
@@ -19,22 +19,11 @@ package dev.nesk.akkurate.examples.ktor.server.plugins
 
 import io.ktor.http.*
 import io.ktor.server.application.*
-import io.ktor.server.plugins.requestvalidation.*
-import io.ktor.server.plugins.statuspages.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 fun Application.configureRouting() {
-    install(StatusPages) {
-        exception<Throwable> { call, cause ->
-            call.respondText(text = "500: $cause", status = HttpStatusCode.InternalServerError)
-        }
-        exception<RequestValidationException> { call, cause ->
-            call.respond(HttpStatusCode.UnprocessableEntity, cause.reasons)
-        }
-    }
-
     routing {
         post("/books") {
             val book = call.receive<Book>()